Revert "lhd"
This reverts commit 44371603841808f4df5ec41a0454eeeaaf3d7cce.

+++ src/main/java/com/redsun/pojo/User.java
@@ -3,10 +3,10 @@ package com.redsun.pojo;
 public class User {
 
     //用户分4种，操作员(operator)，货主(shipper)，承运人(freighter)，管理员(admin)
-    public static final int TYPE_OPERATOR = 1;
-    public static final int TYPE_SHIPPER = 2;
-    public static final int TYPE_FREIGHTER = 3;
-    public static final int TYPE_ADMIN = 4;
+    public static final Byte TYPE_OPERATOR = 1;
+    public static final Byte TYPE_SHIPPER = 2;
+    public static final Byte TYPE_FREIGHTER = 3;
+    public static final Byte TYPE_ADMIN = 4;
     //8位数字id
     private Integer id;
     private String password;
